Boolean `isHilited` prop becomes integer `volume`
instead of a button either being highlighted or not, an integer value gives us more control over hilights

For shuffle button, it allows us to highlight with increasing urgency
It also allows for intuitive "dimmer" states, that function as the opposite of a hilight.
we can also emulate the same isHilited behavoir by checking whether volume is nonzero or not.